fix: update package version number
The version number was omitted in the previous commit by mistake. This commit updates the version number. Because a release for version 2.0.2 was already created and published, while the package version was set to `2.0.1`, version 2.0.2 can no longer be used, because it's tied to that commit. For this reason, this commit skips version 2.0.2 and goes straight to 2.0.3.