feat: return 403 if Learning Assistant feature toggle or instructor toggle disabled
This commit modifies the Learning Assistant access logic.

Previously, the chat view would return a 403 response code if the learning_assistant_available function returned False (i.e. the COURSEWARE_LEARNING_ASSISTANT CourseWaffleFlag was disabled).

Now, the chat view will return a 403 response code if either the learning_assistant_available function returns False or if there is a False override in the LearningAssistantCourseEnabled model corresponding to the course in which the chat is being requested.