feat: add GET endpoint to retrieve whether Learning Assistant is enabled
This commit adds a new GET endpoint to retrieve whether the Learning Assistant feature is enabled in a course, given a course ID. This endpoint will introspect all data that is relevant to determining whether the feature is enabled (i.e. the courseware.learning_assistant CourseWaffleFlag and the LearningAssistantCourseEnabled model).

This endpoint was added so that the frontend can determine whether or not to show the Learning Assistant feature without requiring that the edx-platform read from the LearningAssistantCourseEnabled model in this plugin.