feat: add audit_trial_length attribute to ChatSummaryView response
In order for the frontend to know the currently configured trial length of an audit trial in days, an audit_trial_length attribute is added to the ChatSummaryView response.

This property is supplied whether or not a learner has an audit trial. It represents the number of days in an audit trial as currently configured. It does not represent the number of days in the learner's current trial.