feat: add expiration_date to LearningAssistantAuditTrial model
This commit adds an `expiration_date` field to the `LearningAssistantAuditTrial` model. This is an enhancement that makes it easier to know the exact expiration date of an audit trial. The audit trial length may change over time. For example, the setting may be changed, or we may experiment with different audit trial lengths. In those cases, it would challenging to know what the value of the length of the audit trial was when the learner created their audit trial. Adding this value to the model addresses this issue.