feat: modify system prompt trimming logic and add user to raw_id_fields for LearningAssistantMessageAdmin form. (#173)
* perf: add user to raw_id_fields for LearningAssistantMessageAdmin form

This commit adds the user field of the LearningAssistantMessage model to the raw_id_fields of the LearningAssistantAdmin form to improve performance of the form.

* fix: trim down the unit content to character limit and not the entire system prompt

This commit fixes a bug in the prompt trimming logic. Trimming the entire system prompt to the character limit will result in us trimming the actual static prompt text. What we want, instead, is to trim the unit content, which is the actual dynamic part of the system prompt.